cc26x0: refactor recharge code into setup
All the converted code using corrode resides in setup
with appropriate pointers, move this into there as well.

The recharge code is necessary during long sleep times.
